+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]: https://github.com/nasa/harmony-autotester
 
+### Added:
+
+- TRT-627 - Implemented workflow to retrieve all Harmony services from CMR GraphQL
+  along with all associated collections.
+
 TODO
 
 [unreleased]: https://github.com/nasa/harmony-autotester/